fix bug that was preventing blog content updating

+++ server/controllers/blog.js
@@ -53,13 +53,19 @@ blogRouter.post('/', async (request, response) => {
 // update a blog
 blogRouter.put('/:id', (request, response) => {
     const body = request.body
+    const token = getTokenFrom(request)
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!token || !decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
 
     const blog = {
         title: body.title,
         author: body.author,
-        likes: body.likes
+        likes: body.likes,
+        content: body.content
     }
-  
+    
     Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
         .then(updatedBlog => {
             response.json(updatedBlog.toJSON())
@@ -68,6 +74,12 @@ blogRouter.put('/:id', (request, response) => {
 
 // delete a blog
 blogRouter.delete('/:id', async (request, response) => {
+    const token = getTokenFrom(request)
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!token || !decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
 })
